make fgastify config generic
