+++ src/bin/pohttp-server.ts
@@ -1,6 +1,7 @@
 // tslint:disable-next-line:no-var-requires
 require('make-promises-safe');
 
-import { runServer } from '../services/pohttp/server';
+import { runFastify } from '../services/fastifyUtils';
+import { makeServer } from '../services/pohttp/server';
 
-runServer();
+makeServer().then(runFastify);
